added update thunks to all slices
