+++ src/app/homeSlice.js
@@ -32,8 +32,6 @@ export const updateHomeDataThunk = (data) => {
   return async (dispatch, getState) => {
     const homeData = data;
 
-    console.log(homeData);
-  
     const requestOptions = {
       method: "PUT",
       headers: { "Content-Type": "application/json" },
@@ -58,45 +56,5 @@ export const updateHomeDataThunk = (data) => {
       });
   }
 }
-
-export const updateHomeDataT = async (dispatch, getState) => {
-  let state = getState();
-
-  const homeData = state.home.homeData;
-
-  console.log(homeData);
-
-  const requestOptions = {
-    method: "PUT",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify(homeData),
-  };
-
-  // const response = await fetch(API_URL);
-  // const responseJson = await response.json();
-  // console.log(responseJson);
-
-  fetch(API_URL, requestOptions)
-    .then(async (response) => {
-      const data = await response.json();
-
-      // check for error response
-      if (!response.ok) {
-        // get error message from body or default to response status
-        const error = (data && data.message) || response.status;
-        console.log(error);
-        return Promise.reject(error);
-      }
-
-
-      // setPostId(data.id);
-    })
-    .catch((error) => {
-      // setErrorMessage(error);
-      console.error("There was an error!", error);
-    });
-};
-
-
 
 export default homeSlice.reducer;
