+++ src/app/profileSlice.js
@@ -59,3 +59,5 @@ export const fetchDataThunk = async (dispatch, getState) => {
     })
   );
 };
+
+
